Move sanitisation to extensions
