+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.API/Controllers/BaseController.cs
@@ -173,63 +173,5 @@ namespace UKHO.ExchangeSetService.API.Controllers
                 httpContextAccessor.HttpContext.Response.Headers.Append(LastModifiedDateHeaderKey, model.LastModified.ToString());
             return new StatusCodeResult(StatusCodes.Status304NotModified);
         }
-
-        protected string[] SanitizeProductIdentifiers(string[] productIdentifiers)
-        {
-            if (productIdentifiers == null)
-            {
-                return null;
-            }
-
-            if (productIdentifiers.Any(x => x == null))
-            {
-                return [null];
-            }
-
-            var sanitizedIdentifiers = new List<string>();
-
-            if (productIdentifiers.Length > 0)
-            {
-                foreach (var identifier in productIdentifiers)
-                {
-                    var sanitizedIdentifier = identifier.Trim();
-                    sanitizedIdentifiers.Add(sanitizedIdentifier);
-                }
-            }
-
-            return [.. sanitizedIdentifiers];
-        }
-
-        protected string SanitizeLayoutString(string input)
-        {
-            if (string.IsNullOrWhiteSpace(input))
-            {
-                return ExchangeSetLayout.standard.ToString();
-            }
-
-            // Try parse enum (case-insensitive); fallback to "standard"
-            if (Enum.TryParse<ExchangeSetLayout>(input.Trim(), ignoreCase: true, out var parsed))
-            {
-                return parsed.ToString();
-            }
-
-            return ExchangeSetLayout.standard.ToString();
-        }
-
-        protected string SanitizeStandardString(string input)
-        {
-            if (string.IsNullOrWhiteSpace(input))
-            {
-                return ExchangeSetStandard.s63.ToString();
-            }
-
-            // Try parse enum (case-insensitive); fallback to "s63"
-            if (Enum.TryParse<ExchangeSetStandard>(input.Trim(), ignoreCase: true, out var parsed))
-            {
-                return parsed.ToString();
-            }
-
-            return ExchangeSetStandard.s63.ToString();
-        }
     }
 }
