+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.API/Controllers/ProductBuilderController.cs
@@ -75,9 +75,9 @@ namespace UKHO.ExchangeSetService.API.Controllers
         [SwaggerResponse(statusCode: (int)HttpStatusCode.InternalServerError, type: typeof(InternalServerError), description: "Internal Server Error.")]
         public virtual Task<IActionResult> PostProductIdentifiers([FromBody] string[] productIdentifiers, [FromQuery] string callbackUri, [FromQuery] string exchangeSetStandard, [FromQuery] string exchangeSetLayout)
         {
-            exchangeSetStandard = SanitizeStandardString(exchangeSetStandard);
-            exchangeSetLayout = SanitizeLayoutString(exchangeSetLayout);
-            productIdentifiers = SanitizeProductIdentifiers(productIdentifiers);
+            exchangeSetStandard = exchangeSetStandard.SanitizeExchangeSetStandard();
+            exchangeSetLayout = exchangeSetLayout.SanitizeExchangeSetLayout();
+            productIdentifiers = productIdentifiers.SanitizeProductIdentifiers();
 
             return Logger.LogStartEndAndElapsedTimeAsync(EventIds.ESSPostProductIdentifiersRequestStart, EventIds.ESSPostProductIdentifiersRequestCompleted,
                 "Product Identifiers Endpoint request for _X-Correlation-ID:{correlationId}, ExchangeSetStandard:{exchangeSetStandard} and ExchangeSetLayout:{exchangeSetLayout}",
@@ -144,8 +144,8 @@ namespace UKHO.ExchangeSetService.API.Controllers
         [SwaggerResponse(statusCode: (int)HttpStatusCode.InternalServerError, type: typeof(InternalServerError), description: "Internal Server Error.")]
         public virtual Task<IActionResult> PostproductBuilderByProductVersions([FromBody] List<ProductVersionRequest> productVersionsRequest, string callbackUri, [FromQuery] string exchangeSetStandard, [FromQuery] string exchangeSetLayout)
         {
-            exchangeSetStandard = SanitizeStandardString(exchangeSetStandard);
-            exchangeSetLayout = SanitizeLayoutString(exchangeSetLayout);
+            exchangeSetStandard = exchangeSetStandard.SanitizeExchangeSetStandard();
+            exchangeSetLayout = exchangeSetLayout.SanitizeExchangeSetLayout();
 
             return Logger.LogStartEndAndElapsedTimeAsync(EventIds.ESSPostProductVersionsRequestStart, EventIds.ESSPostProductVersionsRequestCompleted,
                 "Product Versions Endpoint request for _X-Correlation-ID:{correlationId}, ExchangeSetStandard:{exchangeSetStandard} and ExchangeSetLayout:{exchangeSetLayout}",
@@ -206,8 +206,8 @@ namespace UKHO.ExchangeSetService.API.Controllers
         public virtual Task<IActionResult> GetproductBuilderSinceDateTime([FromQuery, SwaggerParameter(Required = true), SwaggerSchema(Format = "date-time")] string sinceDateTime,
             [FromQuery] string callbackUri, [FromQuery] string exchangeSetStandard, [FromQuery] string exchangeSetLayout)
         {
-            exchangeSetStandard = SanitizeStandardString(exchangeSetStandard);
-            exchangeSetLayout = SanitizeLayoutString(exchangeSetLayout);
+            exchangeSetStandard = exchangeSetStandard.SanitizeExchangeSetStandard();
+            exchangeSetLayout = exchangeSetLayout.SanitizeExchangeSetLayout();
 
             return Logger.LogStartEndAndElapsedTimeAsync(EventIds.ESSGetProductsFromSpecificDateRequestStart, EventIds.ESSGetProductsFromSpecificDateRequestCompleted,
                 "Product Data SinceDateTime Endpoint request for _X-Correlation-ID:{correlationId}, ExchangeSetStandard:{exchangeSetStandard}, and ExchangeSetLayout:{exchangeSetLayout}",
